Sun, Jan 22, 2023, 9:42 PM +00:00

+++ filter-more/helpers.c
@@ -7,7 +7,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            double k = (image[i][j].rgbtRed + image[i][j].rgbtBlue + image[i][j].rgbtGreen) / 3.0;
+            int k = (int)((image[i][j].rgbtRed + image[i][j].rgbtBlue + image[i][j].rgbtGreen) / 3.0);
             image[i][j].rgbtRed = k;
             image[i][j].rgbtGreen = k;
             image[i][j].rgbtBlue = k;
